[front] refactor: move uidHistory in its own file
